# docs/api/registry.openapi.yaml
openapi: "3.1.0"
info:
  title: Cartogra Registry API
  version: "0.1.0"
  description: |
    Service CRUD, team ownership, SCM connections, and CI contract checks.

    **Authentication:** All endpoints require a valid JWT (cookie or Bearer token issued by the Gateway).
    The Gateway injects `X-Tenant-Id` — clients must not set this header.

    **CI endpoints** (`/ci/*`) accept `X-Cartogra-Api-Key` instead of JWT.

    **Response envelope**:
    - Success: `{ "data": <T>, "traceId": "<32-hex>" }`
    - Error: `{ "error": { "code", "message", "details" }, "traceId": "<32-hex>" }`
    - Header: `X-Trace-Id: <same 32-hex traceId>`

    **Pagination:** List endpoints use `limit` (default 20, max 100) and `offset` (default 0).
    Paginated `data` shape: `{ "items": [...], "total": N, "limit": 20, "offset": 0 }`.
servers:
  - url: http://localhost:8081/api/v1
    description: Local development (direct)
  - url: http://localhost:8080/api/registry/api/v1
    description: Local development (via Gateway)
...
paths:
...
  /teams:
    get:
      summary: List teams for the authenticated tenant
      operationId: listTeams
      tags: [teams]
      security:
        - cookieAuth: []
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/LimitParam"
        - $ref: "#/components/parameters/OffsetParam"
      responses:
        "200":
          description: Paginated list of teams
          headers:
            X-Trace-Id:
              $ref: "#/components/headers/XTraceId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TeamListResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "429":
          $ref: "#/components/responses/TooManyRequests"

    post:
      summary: Create a new team
      operationId: createTeam
      tags: [teams]
      security:
        - cookieAuth: []
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTeamRequest"
      responses:
        "201":
          description: Team created
          headers:
            X-Trace-Id:
              $ref: "#/components/headers/XTraceId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TeamResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "429":
          $ref: "#/components/responses/TooManyRequests"
  /teams/{teamId}:
    get:
      summary: Get team by ID
      operationId: getTeam
      tags: [teams]
      security:
        - cookieAuth: []
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/TeamIdParam"
      responses:
        "200":
          description: Team found
          headers:
            X-Trace-Id:
              $ref: "#/components/headers/XTraceId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TeamResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "429":
          $ref: "#/components/responses/TooManyRequests"

    put:
      summary: Update team
      operationId: updateTeam
      tags: [teams]
      security:
        - cookieAuth: []
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/TeamIdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTeamRequest"
      responses:
        "200":
          description: Team updated
          headers:
            X-Trace-Id:
              $ref: "#/components/headers/XTraceId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TeamResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "429":
          $ref: "#/components/responses/TooManyRequests"

    delete:
      summary: Soft-delete a team
      operationId: deleteTeam
      tags: [teams]
      security:
        - cookieAuth: []
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/TeamIdParam"
      responses:
        "204":
          description: Team soft-deleted
          headers:
            X-Trace-Id:
              $ref: "#/components/headers/XTraceId"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "429":
          $ref: "#/components/responses/TooManyRequests"

  # ── CI ─────────────────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: cartogra_session
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-Cartogra-Api-Key
  headers:
    XTraceId:
      description: OTel trace ID — 32 lowercase hex characters, no dashes
      schema:
        type: string
        pattern: "^[0-9a-f]{32}$"
  parameters:
...
    TeamIdParam:
      name: teamId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    LimitParam:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    OffsetParam:
      name: offset
      in: query
      schema:
        type: integer
        minimum: 0
        default: 0
  schemas:
...
    ApiError:
      description: Standard error envelope
      type: object
      required: [error, traceId]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: object
              additionalProperties: true
        traceId:
          type: string
          pattern: "^[0-9a-f]{32}$"

    # ── Service ───────────────────────────────────────────────────────────────
...
    TeamDto:
      type: object
      required: [id, tenantId, name, createdAt, updatedAt]
      properties:
        id:
          type: string
          format: uuid
        tenantId:
          type: string
          format: uuid
        name:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CreateTeamRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
    UpdateTeamRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
    TeamResponse:
      type: object
      required: [data, traceId]
      properties:
        data:
          $ref: "#/components/schemas/TeamDto"
        traceId:
          type: string
          pattern: "^[0-9a-f]{32}$"
    TeamListResponse:
      type: object
      required: [data, traceId]
      properties:
        data:
          type: object
          required: [items, total, limit, offset]
          properties:
            items:
              type: array
              items:
                $ref: "#/components/schemas/TeamDto"
            total:
              type: integer
              format: int64
            limit:
              type: integer
            offset:
              type: integer
        traceId:
          type: string
          pattern: "^[0-9a-f]{32}$"

    # ── CI ────────────────────────────────────────────────────────────────────
...
  responses:
    BadRequest:
      description: Invalid request payload or parameters
      headers:
        X-Trace-Id:
          $ref: "#/components/headers/XTraceId"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
    Unauthorized:
      description: Missing or invalid authentication
      headers:
        X-Trace-Id:
          $ref: "#/components/headers/XTraceId"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
...
    NotFound:
      description: Resource not found
      headers:
        X-Trace-Id:
          $ref: "#/components/headers/XTraceId"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
...
    TooManyRequests:
      description: Rate limit exceeded — enforced at the gateway
      headers:
        X-Trace-Id:
          $ref: "#/components/headers/XTraceId"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"